Tienda: subcategorías, filtros, precios en listado, fix imagen hero y precio
- redcase-hero: agrega resolveImg para manejar rutas de uploads correctamente
- Product: cast price a integer (Gs. no tiene centavos, evita error de -1 Gs.)
- Subcategorías: migración parent_id, relaciones en ProductCategory, optgroups en formularios
- redcase-products: filtros de categoría tipo catálogo, show_price y per_page por bloque
- Paginación: lang/es/pagination.php (Anterior / Siguiente en español)
- Media picker: abre en 'Subidos', filtros de carpeta para navegar assets

+++ resources/views/public/store/show.blade.php
@@ -65,7 +65,7 @@
                     <h1 class="product-title">{{ $product->name }}</h1>
 
                     @if($product->price)
-                        <div class="product-price">Gs. {{ number_format((float) $product->price, 0, ',', '.') }}</div>
+                        <div class="product-price">Gs. {{ number_format((int) $product->price, 0, ',', '.') }}</div>
                     @endif
 
                     @if($product->description)
